Corrected type of log level
This was needed for MS

+++ Buildings/Resources/src/ThermalZones/EnergyPlus/C-Sources/InputVariableAllocate.c
@@ -64,7 +64,7 @@ void* EnergyPlusInputVariableAllocate(
   int usePrecompiledFMU,
   const char* fmuName,
   const char* buildingsLibraryRoot,
-  const size_t logLevel,
+  const int logLevel,
   void (*SpawnMessage)(const char *string),
   void (*SpawnError)(const char *string),
   void (*SpawnFormatMessage)(const char *string, ...),
