+++ Buildings/Resources/src/ThermalZones/EnergyPlus/C-Sources/InputVariableAllocate.h
@@ -24,7 +24,7 @@ LBNL_EnergyPlus_EXPORT void* EnergyPlusInputVariableAllocate(
   int usePrecompiledFMU,
   const char* fmuName,
   const char* buildingsLibraryRoot,
-  const size_t logLevel,
+  const int logLevel,
   void (*SpawnMessage)(const char *string),
   void (*SpawnError)(const char *string),
   void (*SpawnFormatMessage)(const char *string, ...),
