+++ Buildings/Resources/src/ThermalZones/EnergyPlus/C-Sources/ZoneAllocate.h
@@ -20,7 +20,7 @@ LBNL_EnergyPlus_EXPORT void* EnergyPlusZoneAllocate(
   int usePrecompiledFMU,
   const char* fmuName,
   const char* buildingsLibraryRoot,
-  const size_t logLevel,
+  const int logLevel,
   void (*SpawnMessage)(const char *string),
   void (*SpawnError)(const char *string),
   void (*SpawnFormatMessage)(const char *string, ...),
